backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
py-feat==0.6.2
scipy==1.9.3
opencv-python==4.8.1.78
numpy==1.23.5
pandas==2.0.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
xgboost==2.0.3